rova: power-libperfmgr: Use performance governor for launch hint
Set "performance" as the governor instead of setting
fixed frequencies. In this way the scaling can be done better as the
governor will compute the desired CPU frequency, potentially based off
of the system's needs, which speeds up dramatically the launch even when
using full Gapps and tons of apps.

Signed-off-by: iusmac <iusico.maxim@libero.it>