rova: prop: Fix too aggressive LMK when low swap breaches low watermark
Most apps are killed very quickly for the reason "low watermark is
breached and swap is low". After some investigations, it turned out that
this happens due to 'swap_is_low' being always true, while only 50-60%
of the swap space is actually used.
The 'swap_is_low' is determined on how the free swap space is calculated
by LMK, which in the case of zRAM is not straightforward like swapping
memory pages on disk. There's a little time gap between when the free
swap space is considered by the system as available memory and when zRAM
actually compresses the data in RAM, which may not compress at all if
the data is incompressible.

To prevent LMK from starting mass killing of apps when there's plenty of
swap space, we lower the level of free swap to 8% (the default value is
10%) to give the zRAM and the CPU some time to complete the job.

Signed-off-by: iusmac <iusico.maxim@libero.it>