rova: qcom-caf: audio: Fix send mute command when in-call volume is 0%
Interestingly, Android never ever sets volume value to 0.0 when in-call
volume level is set to 0%. The lowest possible value we can get is 0.01
(or 1%), which, when mapped to driver level, will return the second-last
level instead of 0. As a result, there's still some audio.
The subsequent call to platform_set_voice_volume is supposed to address
that by sending mute command (Voice Rx Device Mute), and TRULY mute the
in-call volume, but this never happens, as we haven't reached the lowest
possible level because the volume value is not exactly 0.0.
To fix this, we'll round to the nearest tenths place. We only have 5
volume levels, so the precision is irrelevant.

---

V AudioPolicyService: AudioCommandThread() adding set voice volume volume 0.010000
V AudioPolicyService: AudioCommandThread() going to sleep
V AudioPolicyService: inserting command: 3 at index 0, num commands 0
V AudioPolicyService: AudioCommandThread() processing set voice volume volume 0.010000

Signed-off-by: iusmac <iusico.maxim@libero.it>